python_celery/requirements.txt
celery
flower
redis
supervisor
python-dotenv
pandas
ipython